+++ src/views/schedule-actions/schedule-action-edit-form/helpers/transform-describe-schedule-response-to-form-data.ts
@@ -1,7 +1,5 @@
 import isNil from 'lodash/isNil';
 
-import { type RetryPolicy } from '@/__generated__/proto-ts/uber/cadence/api/v1/RetryPolicy';
-import { type _uber_cadence_api_v1_ScheduleAction_StartWorkflowAction as StartWorkflowAction } from '@/__generated__/proto-ts/uber/cadence/api/v1/ScheduleAction';
 import {
   SCHEDULE_CATCH_UP_POLICIES,
   SCHEDULE_OVERLAP_POLICIES,
@@ -9,18 +7,15 @@ import {
 import { type DescribeScheduleResponse } from '@/route-handlers/describe-schedule/describe-schedule.types';
 import formatDurationToSeconds from '@/utils/data-formatters/format-duration-to-seconds';
 import formatInputPayload from '@/utils/data-formatters/format-input-payload';
-import formatPayloadMap from '@/utils/data-formatters/format-payload-map';
 import parseGrpcTimestamp from '@/utils/datetime/parse-grpc-timestamp';
 import losslessJsonStringify from '@/utils/lossless-json-stringify';
-import { type RetryPolicyFormFields } from '@/views/shared/retry-policy-fields/schemas/retry-policy-form-schema';
 
-import {
-  type EditScheduleFormData,
-  type EditScheduleFormPrefillValues,
-  type ExhaustiveDefaults,
-} from '../schedule-action-edit-form.types';
+import { type EditScheduleFormPrefillValues } from '../schedule-action-edit-form.types';
 
 import mapCronExpressionToFormFields from './map-cron-expression-to-form-fields';
+import mapMemoToFormDefault from './map-memo-to-form-default';
+import mapRetryPolicyToFormDefaults from './map-retry-policy-to-form-defaults';
+import mapSearchAttributesToFormDefaults from './map-search-attributes-to-form-defaults';
 
 /**
  * Builds the edit form's default values from an existing schedule.
@@ -85,87 +80,4 @@ export default function transformDescribeScheduleResponseToFormData(
     ),
     memo: mapMemoToFormDefault(startWorkflow?.memo),
   };
-}
-
-/**
- * Has its own exhaustive return type: the top-level one cannot reach into the
- * nested retry policy object, whose sub-fields are all optional too.
- */
-function mapRetryPolicyToFormDefaults(
-  retryPolicy: RetryPolicy | null | undefined
-): ExhaustiveDefaults<RetryPolicyFormFields> {
-  if (!retryPolicy) {
-    return {
-      enableRetryPolicy: false,
-      limitRetries: undefined,
-      retryPolicy: undefined,
-    };
-  }
-
-  const expirationIntervalSeconds = formatDurationToSeconds(
-    retryPolicy.expirationInterval
-  );
-
-  const retryPolicyValues: ExhaustiveDefaults<
-    NonNullable<RetryPolicyFormFields['retryPolicy']>
-  > = {
-    initialIntervalSeconds: stringifyPositiveNumber(
-      formatDurationToSeconds(retryPolicy.initialInterval)
-    ),
-    backoffCoefficient: stringifyPositiveNumber(retryPolicy.backoffCoefficient),
-    maximumIntervalSeconds: stringifyPositiveNumber(
-      formatDurationToSeconds(retryPolicy.maximumInterval)
-    ),
-    maximumAttempts: stringifyPositiveNumber(retryPolicy.maximumAttempts),
-    expirationIntervalSeconds: stringifyPositiveNumber(
-      expirationIntervalSeconds
-    ),
-  };
-
-  return {
-    enableRetryPolicy: true,
-    // The form offers the two limits as an either/or, and the schedule only
-    // carries an expiration interval when it was configured by duration.
-    limitRetries: expirationIntervalSeconds ? 'DURATION' : 'ATTEMPTS',
-    retryPolicy: retryPolicyValues,
-  };
-}
-
-function mapSearchAttributesToFormDefaults(
-  searchAttributes: StartWorkflowAction['searchAttributes'] | undefined
-): EditScheduleFormData['searchAttributes'] {
-  const decoded = formatPayloadMap(searchAttributes ?? null, 'indexedFields')
-    ?.indexedFields as Record<string, unknown> | undefined;
-
-  if (!decoded || Object.keys(decoded).length === 0) {
-    return undefined;
-  }
-
-  return Object.entries(decoded).map(([key, value]) => ({
-    key,
-    // The form only edits primitives; anything richer keeps its JSON text.
-    value:
-      typeof value === 'string' ||
-      typeof value === 'number' ||
-      typeof value === 'boolean'
-        ? value
-        : losslessJsonStringify(value),
-  }));
-}
-
-function mapMemoToFormDefault(
-  memo: StartWorkflowAction['memo'] | undefined
-): string | undefined {
-  const decoded = formatPayloadMap(memo ?? null, 'fields')?.fields as
-    | Record<string, unknown>
-    | undefined;
-
-  return decoded && Object.keys(decoded).length > 0
-    ? losslessJsonStringify(decoded, null, 2)
-    : undefined;
-}
-
-/** Zero and missing values both mean "unset" for these form fields. */
-function stringifyPositiveNumber(value: number | null | undefined) {
-  return value ? String(value) : undefined;
 }
